// types/equipe-aps.ts
export type EquipeApsDetalhada = {
  equipeId: string;
  sigla: string | null;
  nome: string | null;
  situacao: string | null;
  codigoCnes: string;
  nomeUnidade: string | null;
  tipoUnidade: string | null;
  endereco: string | null;
  bairro: string | null;
  nomeArea: string | null;
  descricaoSegmento: string | null;
};